pyproject.toml
[tool.black]
    line-length = 99
    target-version = ['py38']
    include = '\.py$'
[tool.isort]
    profile = "black"
    line_length = 99
    combine_as_imports = true
    atomic = true
    combine_star = true
    from_first = true
    force_alphabetical_sort_within_sections = true
    virtual_env = true
    case_sensitive = true
    remove_redundant_aliases = true
    dedup_headings = true
    
[tool.poetry]
name = "meifwa"
version = "3.2.2"
description = "Created for fun, has tons of commands."
authors = ["KayTwenty"]
license = "MIT"

[tool.poetry.dependencies]
python = ">3.8.0,<4.0"
psutil = "^5.8.0"
PyNaCl = "^1.4.0"
aiosqlite = "^0.17.0"
databases = {extras = ["sqlite"], version = "^0.5.2"}
humanize = "^3.11.0"
hentai = "^3.2.8"
colorama = "^0.4.4"
toml = "^0.10.2"
red-lavalink = {git = "https://github.com/PredaaA/Red-Lavalink", branch = "feature/dpy-voiceprotocol"}
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
Pillow = "^8.3.2"
novus = {git = "https://github.com/Voxel-Fox-Ltd/Novus.git"}
jishaku = {git = "https://github.com/6days9weeks/jishaku.git"}
voxelbotutils = {git = "https://github.com/6days9weeks/VoxelBotUtils", branch = "siesta"}
tabulate = "^0.8.9"
discord-together = "1.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"